requirements.txt
torchvision==0.2.1
http://download.pytorch.org/whl/cpu/torch-0.3.1-cp36-cp36m-linux_x86_64.whl
pillow==6.2.1
kafka-python
redis
python-dotenv
